[fix] fixed bad status of stop request for record.
If record is set as manual and nginx is run in multi-processes mode,
record requests may be distributed to a non-publisher.
Previously, if stop request for record was sent to a non-publisher,
ngx_rtmp_record_init was called without checking if record context
was already initialized, that led to a bad status (204) replied.